standardize api calls from cli
