+++ internal/cli/trigger.go
@@ -1,10 +1,7 @@
 package cli
 
 import (
-	"bytes"
 	"fmt"
-	"io/ioutil"
-	"net/http"
 
 	"github.com/spf13/cobra"
 )
@@ -28,22 +25,8 @@ Otherwise, this command will force a full rebuild.
 
 func triggerUpdate(cmd *cobra.Command, args []string) {
 	resource := args[0]
-	url := fmt.Sprintf("http://localhost:%d/api/trigger", webPort)
 	payload := []byte(fmt.Sprintf(`{"manifest_names":[%q]}`, resource))
-	res, err := http.Post(url, "application/json", bytes.NewBuffer(payload))
-	if err != nil {
-		cmdFail(fmt.Errorf("Could not connect to Tilt at %s: %v", url, err))
-	}
 
-	if res.StatusCode != http.StatusOK {
-		body := "<no response body>"
-		b, err := ioutil.ReadAll(res.Body)
-		if err != nil {
-			cmdFail(fmt.Errorf("Error reading response body from %s: %v", url, err))
-		}
-		if string(b) != "" {
-			body = string(b)
-		}
-		cmdFail(fmt.Errorf("Request to %s failed with status %q: %s", url, res.Status, body))
-	}
+	body := apiPostJson(webPort, "trigger", payload)
+	_ = body.Close()
 }
